pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
advent-of-code-data = "^2.0.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
